Move rendering logic to external module
